Player
Player stam working
CanPunch working

+++ app/src/main/java/com/drew/teamawesomegame/Enemy.java
@@ -4,8 +4,8 @@ public class Enemy {
 
     String enemyName;
     //int enemyNumber;
-    public static int health;
-    public static int maxHealth;
+    public static float health;
+    public static float maxHealth;
     int damage;
     int coinReward;
     int expReward;
